DeleteCommand Class is now abstract, DeleteCaseCommand added
