+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -1,49 +1,32 @@
 package seedu.address.logic.commands;
 
-import static java.util.Objects.requireNonNull;
 
-import java.util.List;
-
-import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.investigationcase.Case;
 
 /**
- * Deletes a case identified using it's displayed index from PIVOT.
+ * Abstract class for deleting different types of items based off of index.
  */
-public class DeleteCommand extends Command {
+public abstract class DeleteCommand extends Command {
 
     public static final String COMMAND_WORD = "delete";
 
+    //TODO: Consider abstracting this out further, find a way to represent any item type or show
+    // of all different item types and the delete command usage
     public static final String MESSAGE_USAGE = COMMAND_WORD
             + ": Deletes the specified type of item identified by the index number"
             + " used in the displayed list.\n"
             + "Parameters: INDEX (must be a positive integer)\n"
             + "Example: " + COMMAND_WORD + " case 1";
 
-    public static final String MESSAGE_DELETE_CASE_SUCCESS = "Deleted Case: %1$s";
+    protected final Index targetIndex;
 
-    private final Index targetIndex;
-
-    public DeleteCommand(Index targetIndex) {
+    protected DeleteCommand(Index targetIndex) {
         this.targetIndex = targetIndex;
     }
 
-    @Override
-    public CommandResult execute(Model model) throws CommandException {
-        requireNonNull(model);
-        List<Case> lastShownList = model.getFilteredCaseList();
-
-        if (targetIndex.getZeroBased() >= lastShownList.size()) {
-            throw new CommandException(Messages.MESSAGE_INVALID_CASE_DISPLAYED_INDEX);
-        }
-
-        Case caseToDelete = lastShownList.get(targetIndex.getZeroBased());
-        model.deleteCase(caseToDelete);
-        return new CommandResult(String.format(MESSAGE_DELETE_CASE_SUCCESS, caseToDelete));
-    }
+    public abstract CommandResult execute(Model model) throws CommandException;
 
     @Override
     public boolean equals(Object other) {
